apply notfoundexception in events.controller

+++ src/events/events.controller.ts
@@ -4,6 +4,7 @@ import {
   Delete,
   Get,
   HttpCode,
+  NotFoundException,
   Param,
   ParseIntPipe,
   Patch,
@@ -29,7 +30,13 @@ export class EventsController {
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
-    return await this.repository.findOneBy({ id });
+    const event = await this.repository.findOneBy({ id });
+
+    if(!event) {
+      throw new NotFoundException();
+    }
+
+    return event;
   }
 
   @Post()
@@ -42,7 +49,11 @@ export class EventsController {
 
   @Patch(':id')
   async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOneBy({ id });
+
+    if(!event) {
+      throw new NotFoundException();
+    }
 
     return await this.repository.save({
       ...event,
@@ -54,7 +65,12 @@ export class EventsController {
   @Delete(':id')
   @HttpCode(204) // No content
   async remove(@Param('id') id) {
-    const event = await this.repository.findOne(id);
+    const event = await this.repository.findOneBy({ id });
+
+    if(!event) {
+      throw new NotFoundException();
+    }
+    
     await this.repository.delete(event);
   }
 }
